vcas

+++ core/record.h
@@ -14,13 +14,13 @@ struct Record {
   std::string topic;
   std::string key;
   std::string message;
-  unsigned time;
+  unsigned long time;
 
   Record(const std::string &topic_, const std::string &message_,
-         const unsigned time_ = 0);
+         const unsigned long time_ = 0);
 
   Record(const std::string &topic_, const std::string &key_,
-         const std::string &message_, const unsigned time_ = 0);
+         const std::string &message_, const unsigned long time_ = 0);
 };
 
 }  // namespace budlab::lib::msg
